Move code generation helpers into test module
